Refactored huge server file into controllers
Refactored flask routes into user and file Flask blueprints.
